requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
iniconfig==2.1.0
joblib==1.5.1
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.4.0
